+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Link to import log in mail ([#12](https://github.com/scm-manager/scm-mail-plugin/pull/12))
+
 ## 2.4.0 - 2020-12-04
 ### Added
 - Send mail for successful or failed repository import ([#9](https://github.com/scm-manager/scm-mail-plugin/pull/9))